requirements.txt
aiohttp==3.7.4.post0
alembic==1.6.5
amqp==5.0.6
async-timeout==3.0.1
attrs==21.2.0
billiard==3.6.4.0
blinker==1.4
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.26.0
Faker==8.10.1
Flask==2.0.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.2
marshmallow-sqlalchemy==0.26.1
multidict==5.1.0
prompt-toolkit==3.0.19
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.21
text-unidecode==1.3
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
yarl==1.6.3
